+++ filesystem/android/src/main/java/com/capacitorjs/plugins/filesystem/FilesystemPlugin.java
@@ -4,6 +4,7 @@ import android.Manifest;
 import android.media.MediaScannerConnection;
 import android.net.Uri;
 import android.os.Build;
+import android.os.Environment;
 import com.capacitorjs.plugins.filesystem.exceptions.CopyFailedException;
 import com.capacitorjs.plugins.filesystem.exceptions.DirectoryExistsException;
 import com.capacitorjs.plugins.filesystem.exceptions.DirectoryNotFoundException;
@@ -17,7 +18,10 @@ import com.getcapacitor.PluginMethod;
 import com.getcapacitor.annotation.CapacitorPlugin;
 import com.getcapacitor.annotation.Permission;
 import com.getcapacitor.annotation.PermissionCallback;
-import java.io.*;
+import com.getcapacitor.plugin.util.HttpRequestHandler;
+import java.io.File;
+import java.io.FileNotFoundException;
+import java.io.IOException;
 import java.nio.charset.Charset;
 import java.nio.file.Files;
 import java.nio.file.attribute.BasicFileAttributes;
@@ -34,6 +38,7 @@ import org.json.JSONException;
 )
 public class FilesystemPlugin extends Plugin {
 
+    static final String PUBLIC_STORAGE = "publicStorage";
     private Filesystem implementation;
 
     @Override
@@ -376,6 +381,35 @@ public class FilesystemPlugin extends Plugin {
         this._copy(call, false);
     }
 
+    @PluginMethod
+    public void downloadFile(PluginCall call) {
+        try {
+            String directory = call.getString("directory", Environment.DIRECTORY_DOWNLOADS);
+
+            if (isPublicDirectory(directory) && !isStoragePermissionGranted()) {
+                requestAllPermissions(call, "permissionCallback");
+            } else {
+                HttpRequestHandler.ProgressEmitter emitter = (bytes, contentLength) -> {
+                    JSObject ret = new JSObject();
+                    ret.put("url", call.getString("url"));
+                    ret.put("bytes", bytes);
+                    ret.put("contentLength", contentLength);
+
+                    notifyListeners("progress", ret);
+                };
+
+                JSObject response = implementation.downloadFile(call, bridge, emitter);
+                // update mediaStore index only if file was written to external storage
+                if (isPublicDirectory(directory)) {
+                    MediaScannerConnection.scanFile(getContext(), new String[] { response.getString("path") }, null, null);
+                }
+                call.resolve(response);
+            }
+        } catch (Exception ex) {
+            call.reject("Error downloading file: " + ex.getLocalizedMessage(), ex);
+        }
+    }
+
     private void _copy(PluginCall call, Boolean doRename) {
         String from = call.getString("from");
         String to = call.getString("to");
@@ -405,6 +439,28 @@ public class FilesystemPlugin extends Plugin {
             call.reject(ex.getMessage());
         } catch (IOException ex) {
             call.reject("Unable to perform action: " + ex.getLocalizedMessage());
+        }
+    }
+
+    @PluginMethod
+    public void checkPermissions(PluginCall call) {
+        if (isStoragePermissionGranted()) {
+            JSObject permissionsResultJSON = new JSObject();
+            permissionsResultJSON.put(PUBLIC_STORAGE, "granted");
+            call.resolve(permissionsResultJSON);
+        } else {
+            super.checkPermissions(call);
+        }
+    }
+
+    @PluginMethod
+    public void requestPermissions(PluginCall call) {
+        if (isStoragePermissionGranted()) {
+            JSObject permissionsResultJSON = new JSObject();
+            permissionsResultJSON.put(PUBLIC_STORAGE, "granted");
+            call.resolve(permissionsResultJSON);
+        } else {
+            requestPermissionForAlias(PUBLIC_STORAGE, call, "permissionCallback");
         }
     }
 
@@ -448,15 +504,19 @@ public class FilesystemPlugin extends Plugin {
             case "stat":
                 stat(call);
                 break;
+            case "downloadFile":
+                downloadFile(call);
+                break;
         }
     }
 
     /**
      * Checks the the given permission is granted or not
-     * @return Returns true if the permission is granted and false if it is denied.
+     * @return Returns true if the app is running on Android 30 or newer or if the permission is already granted
+     * or false if it is denied.
      */
     private boolean isStoragePermissionGranted() {
-        return getPermissionState("publicStorage") == PermissionState.GRANTED;
+        return Build.VERSION.SDK_INT >= Build.VERSION_CODES.R || getPermissionState(PUBLIC_STORAGE) == PermissionState.GRANTED;
     }
 
     /**