+++ device/CHANGELOG.md
@@ -3,6 +3,60 @@
 All notable changes to this project will be documented in this file.
 See [Conventional Commits](https://conventionalcommits.org) for commit guidelines.
 
+## [5.0.6](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/device@5.0.5...@capacitor/device@5.0.6) (2023-07-12)
+
+**Note:** Version bump only for package @capacitor/device
+
+
+
+
+
+## [5.0.5](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/device@5.0.4...@capacitor/device@5.0.5) (2023-06-29)
+
+**Note:** Version bump only for package @capacitor/device
+
+
+
+
+
+## [5.0.4](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/device@5.0.3...@capacitor/device@5.0.4) (2023-06-08)
+
+**Note:** Version bump only for package @capacitor/device
+
+
+
+
+
+## [5.0.3](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/device@5.0.2...@capacitor/device@5.0.3) (2023-06-08)
+
+
+### Bug Fixes
+
+* **device:** webViewVersion not working on Firefox > v108 ([#1419](https://github.com/ionic-team/capacitor-plugins/issues/1419)) ([85cd516](https://github.com/ionic-team/capacitor-plugins/commit/85cd516f06806012350a718b6cd053f9efe583c4))
+
+
+
+
+
+## [5.0.2](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/device@5.0.1...@capacitor/device@5.0.2) (2023-05-09)
+
+**Note:** Version bump only for package @capacitor/device
+
+
+
+
+
+## [5.0.1](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/device@5.0.0...@capacitor/device@5.0.1) (2023-05-05)
+
+
+### Bug Fixes
+
+* Use Capacitor 5 final ([#1574](https://github.com/ionic-team/capacitor-plugins/issues/1574)) ([139c18b](https://github.com/ionic-team/capacitor-plugins/commit/139c18b86a11d31246e952d1a74335ff8ce5dbc2))
+
+
+
+
+
 # [5.0.0](https://github.com/ionic-team/capacitor-plugins/compare/@capacitor/device@5.0.0-beta.1...@capacitor/device@5.0.0) (2023-05-03)
 
 **Note:** Version bump only for package @capacitor/device